+++ ElvUI/Core/StaticPopups.lua
@@ -12,6 +12,8 @@ local format, error = format, error
 
 local CreateFrame = CreateFrame
 local IsAddOnLoaded = IsAddOnLoaded
+local PickupContainerItem = PickupContainerItem
+local DeleteCursorItem = DeleteCursorItem
 local UnitIsDeadOrGhost, InCinematic = UnitIsDeadOrGhost, InCinematic
 local GetBindingFromClick, RunBinding = GetBindingFromClick, RunBinding
 local PurchaseSlot, GetBankSlotCost = PurchaseSlot, GetBankSlotCost
@@ -288,6 +290,29 @@ E.PopupDialogs.RESET_NP_AF = {
 	end,
 	whileDead = 1,
 	hideOnEscape = false,
+}
+
+E.PopupDialogs.DELETE_GRAYS = {
+	text = format('|cffff0000%s|r', L["Delete gray items?"]),
+	button1 = YES,
+	button2 = NO,
+	OnAccept = function()
+		Bags:VendorGrays(true)
+
+		for _, info in ipairs(Bags.SellFrame.Info.itemList) do
+			PickupContainerItem(info[1], info[2])
+			DeleteCursorItem()
+		end
+
+		wipe(Bags.SellFrame.Info.itemList)
+	end,
+	OnShow = function(self)
+		MoneyFrame_Update(self.moneyFrame, E.PopupDialogs.DELETE_GRAYS.Money)
+	end,
+	timeout = 4,
+	whileDead = 1,
+	hideOnEscape = false,
+	hasMoneyFrame = 1,
 }
 
 E.PopupDialogs.BUY_BANK_SLOT = {