+++ ElvUI/Locales/esMX.lua
@@ -341,6 +341,7 @@ L["TopWidget"] = true
 L["PowerBarWidget"] = true
 L["Unhittable:"] = "Imbatible:"
 L["Vehicle Seat Frame"] = "Marco del Asiento del Vehículo"
+L["Vendor / Delete Grays"] = "Vender / Eliminar Grises"
 L["Vendor Grays"] = "Vender Objetos Grises"
 L["Vendoring Grays"] = "Vendiendo Grises"
 L["Voice Overlay"] = "Voz Superpuesta"