+++ ElvUI/Modules/Skins/Blizzard/Tooltip.lua
@@ -7,9 +7,32 @@ local ipairs = ipairs
 local GameTooltip = _G.GameTooltip
 local GameTooltipStatusBar = _G.GameTooltipStatusBar
 
+function S:StyleTooltips()
+	if not (E.private.skins.blizzard.enable and E.private.skins.blizzard.tooltip) then return end
+
+	for _, tt in pairs({
+		_G.ItemRefTooltip,
+		_G.ItemRefShoppingTooltip1,
+		_G.ItemRefShoppingTooltip2,
+		_G.FriendsTooltip,
+		_G.EmbeddedItemTooltip,
+		_G.ReputationParagonTooltip,
+		_G.GameTooltip,
+		_G.ShoppingTooltip1,
+		_G.ShoppingTooltip2,
+		_G.QuickKeybindTooltip,
+		-- ours
+		_G.ElvUIConfigTooltip,
+		_G.ElvUISpellBookTooltip
+	}) do
+		TT:SetStyle(tt)
+	end
+end
+
 function S:TooltipFrames()
 	if not (E.private.skins.blizzard.enable and E.private.skins.blizzard.tooltip) then return end
 
+	S:StyleTooltips()
 	S:HandleCloseButton(_G.ItemRefCloseButton)
 
 	-- Skin Blizzard Tooltips